requirements.txt
Flask==2.0.1
Werkzeug==2.0.2
tensorflow
mysql-connector-python

